Add Path to Next Level

+++ Assets/Gameplay/scripts/Util/Score.cs
@@ -4,5 +4,5 @@ public class Score : MonoBehaviour
 {
     public static float killPoints;
     public static float piecePoints;
-    public static float time;
+    public static int time;
 }
